eliminated version.bash in favor of env variables in setup.sh, don't reinstall n if it's already there

+++ setup.sh
@@ -15,23 +15,16 @@ elif [[ ! $VIRTUALIZE_NODE_VERSION ]]; then
     VIRTUALIZE_NODE_VERSION=$DEFAULT_NODE_VERSION
 fi
 
-#node_version=14.16
-# Retrieve node version to be installed/activated
-source "$VIRTUALIZE_NODE_DIR/version.bash"
-INSTALL_VERSION=$VERSION
-if [[ $1 ]]; then
-    INSTALL_VERSION="$1"
+if [[ ! -d $VIRTUALIZE_NODE_DIR/n ]]; then
+   git clone https://github.com/tj/n $VIRTUALIZE_NODE_DIR/n
+   (
+       cd $VIRTUALIZE_NODE_DIR/n
+       export PREFIX=$VIRTUALIZE_NODE_DIR/node
+       make
+       make install
+       rm -rf $VIRTUALIZE_NODE_DIR/n
+   )
 fi
-
-
-git clone https://github.com/tj/n $VIRTUALIZE_NODE_DIR/n
-(
-    cd $VIRTUALIZE_NODE_DIR/n
-    export PREFIX=$VIRTUALIZE_NODE_DIR/node
-    make
-    make install
-    rm -rf $VIRTUALIZE_NODE_DIR/n
-)
 
 N_PREFIX=$VIRTUALIZE_NODE_DIR/node $VIRTUALIZE_NODE_DIR/node/bin/n $VIRTUALIZE_NODE_VERSION
 export PATH="$VIRTUALIZE_NODE_DIR/node/bin:$PATH"  # apparently npm can't find it's own node install?
